requirements.txt
python-socketio
flask
python-rtmidi
wsgiserver
